Implement getIngredients pagination; Small refactor; Implement updatePlan;

+++ src/main/java/com/solutionscrafted/nutriplanner/dto/PlanRequestDto.java
@@ -3,7 +3,7 @@ package com.solutionscrafted.nutriplanner.dto;
 public record PlanRequestDto(
         String title,
         String description,
-        Integer totalCalories,
+        Double totalCalories,
         Integer numberOfDays,
         Integer clientId,
         String excludeTag) {
